feat: TE7-T123 Action Controller Overview
Create the user controller and add custom events for login and logout
Set session on successful login and end session on logout
Create a car controller with all possible actions(index, new, etc)
Create a new action called search and use it to find a matching car name
Use strong params inside controller actions to see what error arises if the strong parameter is not used, add SS of error.
The user should be able to access the cars list if logged in
Redirect users to the sign-in page if the user is not logged in.
Use filters on some actions with only/except
Use flash messages to notify users about success/failure response
The user should be able to download the pdf file. Put some content into that pdf file.
Practice setting and getting cookies
Use various authentications for some actions for the controller: http_basic_authenticate_with
Use  rescue_from to show the message "Record is not found" if an exception  ActiveRecord::RecordNotFound occurs